- fixed nasty bug with streaming (DEBUG prints still left inside)

+++ branches/refactor_v2_0/audiosources/OGG_Source.cpp
@@ -38,7 +38,7 @@ namespace xal
 				this->samplingRate = info->rate;
 				this->bitsPerSample = 16; // always 16 bit data
 				int bytes = this->bitsPerSample / 8;
-				this->size = (unsigned long)ov_pcm_total(&this->oggStream, -1) * this->channels * bytes;
+				this->size = (int)ov_pcm_total(&this->oggStream, -1) * this->channels * bytes;
 				this->duration = ((float)this->size) / (this->samplingRate * this->channels * bytes);
 			}
 			else
